manomano_interview/requirements.txt
pandas==1.4.2
ruamel-yaml==0.17.21
scikit_learn==1.1.1
tqdm==4.64.0
xgboost==1.6.0
